Workaround for Qt 5.15 and later
Split build process in two stages. The first stage you have to do with
Qt5.9 or below. This will generate the bindings from the Qt5.9 our
earlier header files.

The second stage you can compile with a version of Qt5.15 or later.

The end result are binding plugins that work with QT5.15 QtScript.
However the exposed Qt bindings only reflect Qt5.9 (or the version that
you have used). So no new functionality from Qt5.10 or up is bound.